pyproject.toml
[project]
name = "hpc-sync"
version = "0.1.0"
description = "A tool to sync local files with HPC server using rsync"
authors = [{name = "Charlie Grimshaw", email = "cgrimshaw1@sheffield.ac.uk"}]
dependencies = [
    "hydra-core>=1.3.2",
    "mypy>=1.15.0",
    "rich>=13.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hpc_sync"]

[project.scripts]
hpc-sync = "hpc_sync.cli:main"